renames getData to getDataRequest and moves it to the redux dir
